requirements.txt
django-toolbelt==0.0.1
psycopg2==2.6.1
behave==1.2.5
selenium==2.53.6
